Move Group and Monoid to their own files
