+++ mobius/src/recipes.rs
@@ -223,7 +223,7 @@ mod test {
 
     use core::f64;
 
-    use abstraction::Monoid;
+    use abstraction::monoid::Monoid;
     use test_case::test_case;
 
     use crate::{mobius::MobiusType, nearly::is_nearly};
